Preliminary real time clock implementation
Small cleanups

+++ MainApp/src/Video.cpp
@@ -23,9 +23,9 @@ SDLRenderer::~SDLRenderer()
 // Color correction factors where found by bruteforcing the factors so that the result is somewhat similiar to BGB
 static ggb::RGBA colorCorrection(ggb::RGBA rgb)
 {
-	int r = 0.9 * rgb.r + 0.1 * rgb.g + 0.1 * rgb.b;
-	int g = 0.2 * rgb.r + 0.6 * rgb.g + 0.3 * rgb.b;
-	int b = 0.1 * rgb.r + 0.1 * rgb.g + 0.9 * rgb.b;
+	int r = static_cast<int>(0.9 * rgb.r + 0.1 * rgb.g + 0.1 * rgb.b);
+	int g = static_cast<int>(0.2 * rgb.r + 0.6 * rgb.g + 0.3 * rgb.b);
+	int b = static_cast<int>(0.1 * rgb.r + 0.1 * rgb.g + 0.9 * rgb.b);
 
 	rgb.r = std::clamp(r, 0, 255);
 	rgb.g = std::clamp(g, 0, 255);
